refactor: move form logic to  useInterestsForm composable
